Pokemons Comparison/requirements.txt
streamlit==1.32.0
plotly_express==0.4.1
plotly==5.18.0
st-annotated-text==4.0.1